Started splitting tests
Closes #785
