+++ tests/common/CMakeLists.txt
@@ -1,0 +1,5 @@
+# library common for test utils
+
+add_library(tiglCommonTestUtils INTERFACE)
+
+target_include_directories(tiglCommonTestUtils INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
