+++ tests/integrationtests/TestData/singleModel_withGenericSystems_cube.xml
