+++ tests/integrationtests/TestData/singleModel_withGenericSystems_sphere.xml
