+++ tests/integrationtests/TestData/test_fuselage_walls_ex3_labyrinth.xml
