+++ tests/integrationtests/TestData/test_fuselage_walls_ex4_compartment.xml
