+++ tests/integrationtests/testShapeInvalidation.cpp
