+++ tests/integrationtests/tiglStructuralElements.cpp
